% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bids_data.R
\name{add_bids_data}
\alias{add_bids_data}
\title{Use BIDS structure for a data file}
\usage{
add_bids_data(path, participant_label, session_label = "01",
  data_type = NULL, extension = ".nii.gz", add_session = TRUE,
  modality_label, acquisition_label = NULL, reconstruction_label = NULL,
  run_index = NULL, recursive = TRUE)
}
\arguments{
\item{path}{Directory to put the BIDS data structure folders}

\item{participant_label}{label/identifiers for participants.  The
folders will be prefixed with \code{sub-}.}

\item{session_label}{label/identifiers for sessions.  The
folders will be prefixed with \code{ses-}.}

\item{data_type}{Type of data to be added, will determine the sub-folder}

\item{extension}{Extension of file to be added}

\item{add_session}{Indicator if \code{session_label} should be added
to the filename.}

\item{modality_label}{Label of the imaging modality (e.g. T1w)}

\item{acquisition_label}{Label of acquisition (e.g. highres)}

\item{reconstruction_label}{Label of reconstruction}

\item{run_index}{Index for which run (e.g. 01, 02)}

\item{recursive}{Should the \code{path} be created recursively if not
available}
}
\value{
A \code{data.frame} of paths
}
\description{
Creates necessary sub-folders of a path for a
BIDS structure
}
\note{
The format is based on (without line breaks):
sub-<participant_label>[_ses-<session_label>][_acq-<label>]
[_rec-<label>][_run-<index>]_<modality_label>.<extension>
}
