% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bids_subject.R
\name{add_bids_subject}
\alias{add_bids_subject}
\title{Use BIDS structure}
\usage{
add_bids_subject(path, participant_label, session_label = "01",
  data_type = c("anat", "func", "dwi", "fmap"), recursive = TRUE)
}
\arguments{
\item{path}{Directory to put the BIDS data structure folders}

\item{participant_label}{label/identifiers for participants.  The
folders will be prefixed with \code{sub-}.}

\item{session_label}{label/identifiers for sessions.  The
folders will be prefixed with \code{ses-}.}

\item{data_type}{Type of data that will be added.}

\item{recursive}{Should the \code{path} be created recursively if not
available}
}
\value{
A \code{data.frame} of paths and an indicator
if they were created
}
\description{
Creates necessary sub-folders of a path for a
BIDS structure
}
