% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_alphanmeric.R
\name{is_alphanumeric}
\alias{is_alphanumeric}
\alias{all_alphanumeric}
\alias{check_alphanumeric}
\title{Checks alphanumeric format of a string}
\usage{
is_alphanumeric(x)

all_alphanumeric(x)

check_alphanumeric(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Logical vector (same length as \code{x})
}
\description{
These functions wrap a simple regular expression for
checking if a vector is alpha-numeric characters only.
}
\note{
The functions \code{all_alphanumeric} simply wraps
\code{is_alphanumeric} within \code{all}, and \code{check_alphanumeric}
simply wraps \code{all_alphanumeric} within \code{stop}.
}
\examples{
x = c("hey", "1hey", "hey1", "$hey", "-o", "/", " hey")
is_alphanumeric(x)
}
