% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bids.R
\name{use_bids}
\alias{use_bids}
\title{Use BIDS structure}
\usage{
use_bids(path, recursive = TRUE)
}
\arguments{
\item{path}{Directory to put the BIDS data structure folders}

\item{recursive}{Should the \code{path} be created recursively if not
available}
}
\description{
Creates necessary sub-folders of a path for a
BIDS structure
}
